requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
agno==1.5.4
    # via -r requirements.in
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.16
    # via -r requirements.in
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
asgiref==3.8.1
    # via opentelemetry-instrumentation-asgi
attrs==25.3.0
    # via aiohttp
backoff==2.2.1
    # via posthog
bcrypt==4.3.0
    # via chromadb
build==1.2.2.post1
    # via chromadb
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
charset-normalizer==3.4.1
    # via requests
chroma-hnswlib==0.7.6
    # via chromadb
chromadb==0.6.3
    # via -r requirements.in
click==8.1.8
    # via
    #   typer
    #   uvicorn
coloredlogs==15.0.1
    # via onnxruntime
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-semantic-conventions
distro==1.9.0
    # via
    #   openai
    #   posthog
docstring-parser==0.16
    # via agno
durationpy==0.9
    # via kubernetes
fastapi==0.115.12
    # via
    #   -r requirements.in
    #   chromadb
filelock==3.18.0
    # via huggingface-hub
flatbuffers==25.2.10
    # via onnxruntime
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.3.2
    # via huggingface-hub
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via agno
google-auth==2.39.0
    # via kubernetes
googleapis-common-protos==1.70.0
    # via opentelemetry-exporter-otlp-proto-grpc
greenlet==3.2.0
    # via sqlalchemy
grpcio==1.71.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.8
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   agno
    #   chromadb
    #   ollama
    #   openai
huggingface-hub==0.30.2
    # via tokenizers
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.6.1
    # via opentelemetry-api
importlib-resources==6.5.2
    # via chromadb
iniconfig==2.1.0
    # via pytest
jiter==0.9.0
    # via openai
kubernetes==32.0.1
    # via chromadb
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mmh3==5.1.0
    # via chromadb
monotonic==1.6
    # via posthog
mpmath==1.3.0
    # via sympy
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
numpy==2.2.4
    # via
    #   chroma-hnswlib
    #   chromadb
    #   onnxruntime
    #   pandas
oauthlib==3.2.2
    # via
    #   kubernetes
    #   requests-oauthlib
ollama==0.4.7
    # via -r requirements.in
onnxruntime==1.21.0
    # via chromadb
openai==1.74.0
    # via -r requirements.in
opentelemetry-api==1.32.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-exporter-otlp-proto-common==1.32.0
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.32.0
    # via chromadb
opentelemetry-instrumentation==0.53b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asgi==0.53b0
    # via opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-fastapi==0.53b0
    # via chromadb
opentelemetry-proto==1.32.0
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.32.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions==0.53b0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-sdk
opentelemetry-util-http==0.53b0
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
orjson==3.10.16
    # via chromadb
overrides==7.7.0
    # via chromadb
packaging==24.2
    # via
    #   build
    #   huggingface-hub
    #   onnxruntime
    #   opentelemetry-instrumentation
    #   pytest
pandas==2.2.3
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
posthog==3.24.1
    # via chromadb
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.4
    # via
    #   googleapis-common-protos
    #   onnxruntime
    #   opentelemetry-proto
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.3
    # via
    #   -r requirements.in
    #   agno
    #   chromadb
    #   fastapi
    #   ollama
    #   openai
    #   pydantic-settings
pydantic-core==2.33.1
    # via pydantic
pydantic-settings==2.8.1
    # via agno
pygments==2.19.1
    # via rich
pypika==0.48.9
    # via chromadb
pyproject-hooks==1.2.0
    # via build
pytest==8.1.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   kubernetes
    #   pandas
    #   posthog
python-dotenv==1.1.0
    # via
    #   agno
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via agno
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   agno
    #   chromadb
    #   huggingface-hub
    #   kubernetes
    #   uvicorn
requests==2.32.3
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   kubernetes
    #   posthog
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via kubernetes
rich==14.0.0
    # via
    #   agno
    #   chromadb
    #   typer
rsa==4.9
    # via google-auth
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   kubernetes
    #   posthog
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.38
    # via -r requirements.in
starlette==0.46.2
    # via fastapi
sympy==1.13.1
    # via onnxruntime
tenacity==9.1.2
    # via chromadb
tokenizers==0.21.1
    # via chromadb
tomli==2.2.1
    # via agno
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   chromadb
    #   huggingface-hub
    #   openai
typer==0.15.2
    # via
    #   agno
    #   chromadb
typing-extensions==4.13.2
    # via
    #   agno
    #   anyio
    #   chromadb
    #   fastapi
    #   huggingface-hub
    #   openai
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via
    #   kubernetes
    #   requests
uvicorn[standard]==0.34.1
    # via
    #   -r requirements.in
    #   chromadb
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
websocket-client==1.8.0
    # via kubernetes
// ...existing code...
websockets==15.0
    # via
    #   -r requirements.in
    #   uvicorn
wrapt==1.17.2
    # via
    #   deprecated
    #   opentelemetry-instrumentation
yarl==1.19.0
    # via aiohttp
zipp==3.21.0
    # via importlib-metadata
